Completed initial transition of simulation logic to runners in the emulator
