+++ lightworks/emulator/simulation/probability_distribution.py
@@ -12,13 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from collections.abc import Callable
 from typing import Any
 
 from multimethod import multimethod
 
 from ...sdk.circuit.photonic_compiler import CompiledPhotonicCircuit
 from ...sdk.state import State
-from ..backends.fock_backend import FockBackend
 from ..state import AnnotatedState
 
 
@@ -26,7 +26,7 @@ from ..state import AnnotatedState
 def pdist_calc(
     circuit: CompiledPhotonicCircuit,
     inputs: dict[State, int | float],
-    backend: FockBackend,
+    probability_func: Callable,
 ) -> dict[State, float]:
     """
     Calculate the output state probability distribution for cases where
@@ -41,8 +41,8 @@ def pdist_calc(
         inputs (dict) : The inputs to the system and their associated
             probabilities.
 
-        backend (FockBackend) : A backend object which provides the required
-            methods for calculation of the probability distribution.
+        probability_func (Callable) : A method for calculation of a probability
+            distribution, given a circuit and input state.
 
     Returns:
 
@@ -53,7 +53,7 @@ def pdist_calc(
     # Loop over each possible input
     for istate, prob in inputs.items():
         # Calculate sub distribution
-        sub_dist = backend.full_probability_distribution(circuit, istate)
+        sub_dist = probability_func(circuit, istate)
         if not pdist:
             if prob == 1:
                 pdist = sub_dist
@@ -77,7 +77,7 @@ def pdist_calc(
 def annotated_state_pdist_calc(
     circuit: CompiledPhotonicCircuit,
     inputs: dict[AnnotatedState, int | float],
-    backend: FockBackend,
+    probability_func: Callable,
 ) -> dict[State, float]:
     """
     Perform output state probability distribution calculation using complex
@@ -91,8 +91,8 @@ def annotated_state_pdist_calc(
         inputs (dict) : The inputs to the system and their associated
                         probabilities.
 
-        backend (FockBackend) : A backend object which provides the required
-            methods for calculation of the probability distribution.
+        probability_func (Callable) : A method for calculation of a probability
+            distribution, given a circuit and input state.
 
     Returns:
 
@@ -124,8 +124,8 @@ def annotated_state_pdist_calc(
     unique_results: dict[State, Any] = {}
     for in_state in unique_inputs:
         # Calculate sub distribution and store
-        unique_results[in_state[: circuit.n_modes]] = (
-            backend.full_probability_distribution(circuit, in_state)
+        unique_results[in_state[: circuit.n_modes]] = probability_func(
+            circuit, in_state
         )
 
     # Pre-calculate dictionary items to improve speed
