+++ lightworks/emulator/simulation/runner.py
@@ -12,14 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""
-Script to store various useful functions for the simulation aspect of the code.
-"""
+from abc import ABCMeta, abstractmethod
 
 
-def state_to_string(state: list) -> str:
-    """Converts the provided state to a string with ket notation."""
-    string = "|"
-    for s in state:
-        string += str(s) + ","
-    return string[:-1] + ">"
+class RunnerABC(metaclass=ABCMeta):
+    """
+    Base class for all simulation routines in the emulator.
+    """
+
+    @abstractmethod
+    def run(self) -> dict: ...  # noqa: D102
