+++ lightworks/emulator/utils/__init__.py
@@ -15,4 +15,5 @@
 # ruff: noqa: F401, F403
 
 from .exceptions import *
-from .state_utils import annotated_state_to_string
+from .sim import check_photon_numbers
+from .state_utils import annotated_state_to_string, fock_basis
